Commit 03/12/2023
Includes:

Base Assets (Castle Towers)

Scripts:
-Projectile
-TowerDetectEnemy (Still need to work on priorities a little bit)
-TowerPlacement (WIP)
-TowerTargeting

Removed:

-Utils Script
-Unused Tower Assets
-GameAssets Script
-Cannon AI Script
-All scripts involving the grid in order to implement a better system using a TileMap.

The basic Tower Scripts are nearly done and can be used to add new types and projectiles. At the moment they can be placed in Game Mode but still trying to find a way to lock it onto a grid for the 3 lanes.